"fatcored out more of Flash from main HAL; updates #1173"
