+++ software/controller/lib/hal/flash.h
@@ -19,7 +19,22 @@ limitations under the License.
 #include <cstddef>
 #include <cstdint>
 
-// Flash memory location & size info
-inline constexpr uint32_t FlashStartAddr{0x08000000};
-inline constexpr size_t FlashSize{32 * 1024};
-inline constexpr size_t FlashPageSize{2 * 1024};
+namespace Flash {
+
+void FlashInit();
+
+// Erase a page of flash given the starting address of that page.
+bool FlashErasePage(uint32_t address);
+
+// Write data to flash memory at the specified address
+// Note that the STM32 only allows writing in multiples of 8 bytes
+// to addresses that are a multiple of 8.
+//
+// @param addr - Address in flash memory at which to write.
+//               NOTE - must be a multiple of 8
+// @param data - Pointer to data to write
+// @param ct   - Number of bytes to write
+//               NOTE - must be a multiple of 8
+bool FlashWrite(uint32_t addr, void *data, size_t ct);
+
+}  // namespace Flash
